requirements.txt
# YaTeApruebo - Sistema de Análisis Financiero
# Dependencias principales
python-telegram-bot==20.7
openai==1.52.0
PyPDF2==3.0.1
reportlab==4.0.9
python-dotenv==1.0.0
supabase==2.0.2
requests==2.31.0
pandas==2.1.4
numpy==1.25.2
jinja2==3.1.2
aiofiles==23.2.1
loguru==0.7.2
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
python-dateutil==2.8.2
httpx==0.25.2
python-multipart==0.0.20

